requirements.txt
# Core
numpy>=1.24
torch>=2.1
scipy

# Training utilities
tqdm>=4.66 
pyyaml>=6.0
pandas